fixes: python -m build fails
python -m build fails with:
> running check
> error: The docutils package is needed.

since check.restructuredtext is set to 1 in setup.cfg, setuptools with
require docutils. I think this is needed to parse the structured text
README.
That dependency needs to be specify in pyproject.toml file. It is recommended practice to have a pyproject.toml file anyway.
https://packaging.python.org/en/latest/guides/modernize-setup-py-project/#should-pyproject-toml-be-added